+++ bootique-job-consul/src/main/java/io/bootique/job/consul/ConsulJobConfig.java
@@ -22,7 +22,7 @@ import io.bootique.annotation.BQConfig;
 import io.bootique.annotation.BQConfigProperty;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 @BQConfig("Consul jobs configuration")
 public class ConsulJobConfig {