+++ bootique-job/src/main/java/io/bootique/job/JobRegistry.java
@@ -31,13 +31,13 @@ public interface JobRegistry {
     Job getJob(String jobName);
 
     /**
-     * @deprecated since 0.26 use correctly named {@link #allowsSimultaneousExecutions(String)} method
+     * @deprecated since 1.0.RC1 use correctly named {@link #allowsSimultaneousExecutions(String)} method
      */
     @Deprecated
     boolean allowsSimlutaneousExecutions(String jobName);
 
     /**
-     * @since 0.26
+     * @since 1.0.RC1
      */
     boolean allowsSimultaneousExecutions(String jobName);
 }