+++ bootique-job/src/main/java/io/bootique/job/value/Cron.java
@@ -17,10 +17,7 @@
  * under the License.
  */
 
-/**
- * @since 0.26
- */
-package io.bootique.job.scheduler;
+package io.bootique.job.value;
 
 import java.util.Objects;
 
@@ -39,35 +36,41 @@ import java.util.Objects;
  *     <li>"0 0 0 25 12 ?" - every Christmas Day at midnight</li>
  * </ul>
  *
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class Cron {
 
-    private String expression;
+	private String expression;
 
-    public Cron(String expression) {
-        this.expression = expression;
-    }
+	public Cron(String expression) {
+		this.expression = expression;
+	}
 
-    public String getExpression() {
-        return expression;
-    }
+	public String getExpression() {
+		return expression;
+	}
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Cron cron = (Cron) o;
-        return Objects.equals(expression, cron.expression);
-    }
+	@Override
+	public boolean equals(Object o) {
+		if (this == o) return true;
+		if (o == null || getClass() != o.getClass()) return false;
+		Cron cron = (Cron) o;
+		return Objects.equals(expression, cron.expression);
+	}
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(expression);
-    }
+	@Override
+	public int hashCode() {
+		if (expression == null) {
+			return 0;
+		}
 
-    @Override
-    public String toString() {
-        return expression;
-    }
+		return Objects.hash(expression);
+	}
+
+	@Override
+	public String toString() {
+		return "Cron {" +
+				"expression='" + expression + '\'' +
+				'}';
+	}
 }